Cleanup non-Stable stuff from futures::__rt
