+++ futures-async-runtime/src/stream.rs
@@ -1,47 +1,50 @@
+use std::mem::Pin;
 use std::ops::{Generator, GeneratorState};
-use std::marker::PhantomData;
+use std::marker::{PhantomData, Unpin};
 
-use futures_core::task;
-use futures_core::{Poll, Async, Stream};
+use futures_core::{Poll, Async, task};
+use futures_stable::StableStream;
 
 use super::{CTX, Reset, IsResult};
 
-pub trait MyStream<T, U: IsResult<Ok=()>>: Stream<Item=T, Error=U::Err> {}
+pub trait MyStableStream<T, U: IsResult<Ok=()>>: StableStream<Item=T, Error=U::Err> {}
 
-impl<F, T, U> MyStream<T, U> for F
-    where F: Stream<Item = T, Error = U::Err> + ?Sized,
+impl<F, T, U> MyStableStream<T, U> for F
+    where F: StableStream<Item = T, Error = U::Err> + ?Sized,
           U: IsResult<Ok=()>
 {}
 
 /// Small shim to translate from a generator to a stream.
-struct GenStream<U, T> {
+struct GenStableStream<U, T> {
     gen: T,
     done: bool,
     phantom: PhantomData<U>,
 }
 
-pub fn gen_stream<T, U>(gen: T) -> impl MyStream<U, T::Return>
+impl<U, T> !Unpin for GenStableStream<U, T> { }
+
+pub fn gen_stream<T, U>(gen: T) -> impl MyStableStream<U, T::Return>
     where T: Generator<Yield = Async<U>>,
           T::Return: IsResult<Ok = ()>,
 {
-    GenStream { gen, done: false, phantom: PhantomData }
+    GenStableStream { gen, done: false, phantom: PhantomData }
 }
 
-impl<U, T> Stream for GenStream<U, T>
+impl<U, T> StableStream for GenStableStream<U, T>
     where T: Generator<Yield = Async<U>>,
           T::Return: IsResult<Ok = ()>,
 {
     type Item = U;
     type Error = <T::Return as IsResult>::Err;
 
-    fn poll_next(&mut self, ctx: &mut task::Context) -> Poll<Option<Self::Item>, Self::Error> {
+    fn poll_next(mut self: Pin<Self>, ctx: &mut task::Context) -> Poll<Option<Self::Item>, Self::Error> {
         CTX.with(|cell| {
             let _r = Reset::new(ctx, cell);
-            if self.done { return Ok(Async::Ready(None)) }
-            // Because we are controlling the creation of our underlying
-            // generator, we know that this is definitely a movable generator
-            // so calling resume is always safe.
-            match unsafe { self.gen.resume() } {
+            let this: &mut Self = unsafe { Pin::get_mut(&mut self) };
+            if this.done { return Ok(Async::Ready(None)) }
+            // This is an immovable generator, but since we're only accessing
+            // it via a Pin this is safe.
+            match unsafe { this.gen.resume() } {
                 GeneratorState::Yielded(Async::Ready(e)) => {
                     Ok(Async::Ready(Some(e)))
                 }
@@ -49,7 +52,7 @@ impl<U, T> Stream for GenStream<U, T>
                     Ok(Async::Pending)
                 }
                 GeneratorState::Complete(e) => {
-                    self.done = true;
+                    this.done = true;
                     e.into_result().map(|()| Async::Ready(None))
                 }
             }
